added ChannelStream and StringStream(officially)
