+++ ts_src/utils/index.ts
@@ -1,6 +1,6 @@
 
 export * from './decorators';
-export * from './sstream';
+export * from './streams';
 export * from './subscriber';
 export * from './typedefs';
 export * from './dictionary';
